refactor(codegen): move getCommandOptions to parsers-commons
